+++ runtime/service/src/test/java/org/apache/polaris/service/quarkus/catalog/PolicyCatalogTest.java
@@ -42,8 +42,6 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.function.Supplier;
-import org.apache.commons.lang3.NotImplementedException;
 import org.apache.iceberg.CatalogProperties;
 import org.apache.iceberg.Schema;
 import org.apache.iceberg.catalog.Namespace;
@@ -70,11 +68,7 @@ import org.apache.polaris.core.persistence.MetaStoreManagerFactory;
 import org.apache.polaris.core.persistence.PolarisEntityManager;
 import org.apache.polaris.core.persistence.PolarisMetaStoreManager;
 import org.apache.polaris.core.persistence.PolicyMappingAlreadyExistsException;
-import org.apache.polaris.core.persistence.bootstrap.RootCredentialsSet;
 import org.apache.polaris.core.persistence.cache.InMemoryEntityCache;
-import org.apache.polaris.core.persistence.dao.entity.BaseResult;
-import org.apache.polaris.core.persistence.dao.entity.PrincipalSecretsResult;
-import org.apache.polaris.core.persistence.transactional.TransactionalPersistence;
 import org.apache.polaris.core.policy.PredefinedPolicyTypes;
 import org.apache.polaris.core.policy.exceptions.NoSuchPolicyException;
 import org.apache.polaris.core.policy.exceptions.PolicyInUseException;
@@ -153,7 +147,7 @@ public class PolicyCatalogTest {
       new PolicyAttachmentTarget(
           PolicyAttachmentTarget.TypeEnum.TABLE_LIKE, List.of(TABLE.toString().split("\\.")));
 
-  @Inject MetaStoreManagerFactory managerFactory;
+  @Inject MetaStoreManagerFactory metaStoreManagerFactory;
   @Inject UserSecretsManagerFactory userSecretsManagerFactory;
   @Inject PolarisConfigurationStore configurationStore;
   @Inject PolarisStorageIntegrationProvider storageIntegrationProvider;
@@ -191,12 +185,12 @@ public class PolicyCatalogTest {
                 testInfo.getTestMethod().map(Method::getName).orElse("test"), System.nanoTime());
     RealmContext realmContext = () -> realmName;
     QuarkusMock.installMockForType(realmContext, RealmContext.class);
-    metaStoreManager = managerFactory.getOrCreateMetaStoreManager(realmContext);
+    metaStoreManager = metaStoreManagerFactory.getOrCreateMetaStoreManager(realmContext);
     userSecretsManager = userSecretsManagerFactory.getOrCreateUserSecretsManager(realmContext);
     polarisContext =
         new PolarisCallContext(
             realmContext,
-            managerFactory.getOrCreateSessionSupplier(realmContext).get(),
+            metaStoreManagerFactory.getOrCreateSessionSupplier(realmContext).get(),
             diagServices,
             configurationStore,
             Clock.systemDefaultZone());
@@ -269,9 +263,10 @@ public class PolicyCatalogTest {
             callContext, entityManager, securityContext, CATALOG_NAME);
     TaskExecutor taskExecutor = Mockito.mock();
     RealmEntityManagerFactory realmEntityManagerFactory =
-        new RealmEntityManagerFactory(createMockMetaStoreManagerFactory());
+        new RealmEntityManagerFactory(metaStoreManagerFactory);
     this.fileIOFactory =
-        new DefaultFileIOFactory(realmEntityManagerFactory, managerFactory, configurationStore);
+        new DefaultFileIOFactory(
+            realmEntityManagerFactory, metaStoreManagerFactory, configurationStore);
 
     StsClient stsClient = Mockito.mock(StsClient.class);
     when(stsClient.assumeRole(isA(AssumeRoleRequest.class)))
@@ -310,42 +305,6 @@ public class PolicyCatalogTest {
   @AfterEach
   public void after() throws IOException {
     metaStoreManager.purge(polarisContext);
-  }
-
-  private MetaStoreManagerFactory createMockMetaStoreManagerFactory() {
-    return new MetaStoreManagerFactory() {
-      @Override
-      public PolarisMetaStoreManager getOrCreateMetaStoreManager(RealmContext realmContext) {
-        return metaStoreManager;
-      }
-
-      @Override
-      public Supplier<TransactionalPersistence> getOrCreateSessionSupplier(
-          RealmContext realmContext) {
-        return () -> ((TransactionalPersistence) polarisContext.getMetaStore());
-      }
-
-      @Override
-      public StorageCredentialCache getOrCreateStorageCredentialCache(RealmContext realmContext) {
-        return new StorageCredentialCache(realmContext, configurationStore);
-      }
-
-      @Override
-      public InMemoryEntityCache getOrCreateEntityCache(RealmContext realmContext) {
-        return new InMemoryEntityCache(realmContext, configurationStore, metaStoreManager);
-      }
-
-      @Override
-      public Map<String, PrincipalSecretsResult> bootstrapRealms(
-          Iterable<String> realms, RootCredentialsSet rootCredentialsSet) {
-        throw new NotImplementedException("Bootstrapping realms is not supported");
-      }
-
-      @Override
-      public Map<String, BaseResult> purgeRealms(Iterable<String> realms) {
-        throw new NotImplementedException("Purging realms is not supported");
-      }
-    };
   }
 
   @Test